Changing file reader and updating characteristics for 200 images

+++ Pre-processing/main.m
@@ -1,11 +1,41 @@
-for i = 1:8
-    fileName = ['IMD00' int2str(i) '.bmp'];
-    img = imread(fileName);
+for i = 2:9
+    fileName = ['../Dataset/IMD00' int2str(i) '.bmp'];
+    
+    if exist(fileName, 'file') == 2
+        img = imread(fileName);
 
-    tumor = preProcessing(img);
+        tumor = preProcessing(img);
 
-    folder = 'C:\Users\raulmacintosh\Documents\Projeto\Data extraction\';
-    newimagename = [folder fileName];
-    imwrite(tumor, newimagename);
+        folder = 'C:\Users\raulmacintosh\Documents\Projeto\Data extraction\images\';
+        newimagename = [folder 'IMD00' int2str(i) '.bmp'];
+        imwrite(tumor, newimagename);
+    end
 end
 
+for i = 10:99
+    fileName = ['../Dataset/IMD0' int2str(i) '.bmp'];
+    
+    if exist(fileName, 'file') == 2
+        img = imread(fileName);
+
+        tumor = preProcessing(img);
+
+        folder = 'C:\Users\raulmacintosh\Documents\Projeto\Data extraction\images\';
+        newimagename = [folder 'IMD0' int2str(i) '.bmp'];
+        imwrite(tumor, newimagename);
+    end
+end
+
+for i = 100:437
+    fileName = ['../Dataset/IMD' int2str(i) '.bmp'];
+    
+    if exist(fileName, 'file') == 2
+        img = imread(fileName);
+
+        tumor = preProcessing(img);
+
+        folder = 'C:\Users\raulmacintosh\Documents\Projeto\Data extraction\images\';
+        newimagename = [folder 'IMD' int2str(i) '.bmp'];
+        imwrite(tumor, newimagename);
+    end
+end
